backend/requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14,<3.15
django-cors-headers>=4.3.1,<4.4
Pillow>=10.1,<10.2
celery>=5.3.6,<5.4
redis>=5.0.1,<5.1
django-celery-beat>=2.5.0,<2.6
django-celery-results>=2.5.1,<2.6
python-decouple>=3.8,<3.9
dj-database-url>=2.1.0,<2.2 
psycopg2-binary>=2.9
django-filter>=23.2,<24.0
